scripts/requirements.txt
pyserial>=3.5
Pillow>=10.0.0
matplotlib>=3.5.0
numpy>=1.21.0
qrcode[pil]>=7.4.0
